migrate cur_height from chain storage to blockchain parser
We prefer to maintain the mutable current height state
on the outer `BlockchainParser` struct, making the inner
`ChainStorage` a tiny bit more functional.

As a bonus, this allows interested library users to
employ aribtrary access patterns, for example querying
every nth block, which was previously impossible due
to the imposed sequential nature of using `ChainStorage`.